Fixed Implement Drop Down Commit Issues
The drop-down now is set by the parent view and it is dynamically
set to the appropriate suggestions based on if we are rendering
a tag view for artists or for tags.

The suggestions also remove the tags/artists of the current song so we
are not suggesting the same tags/artists again (helps against duplicates).